Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	renamed:    gen_ai_python/basics/conditional -> gen_ai_python/basics/conditional.py
	modified:   gen_ai_python/matplot/4_pie_plot.py
	modified:   logic_building/1_basic/15_switch_case.c
	modified:   logic_building/1_basic/2_marks_cal.c
	modified:   logic_building/1_basic/7_print_ascii.c
	new file:   logic_building/2_loop/tempCodeRunnerFile.c
	modified:   logic_building/4_array/1_increase_array.c
	modified:   logic_building/6_string_array/5_merge_str.c
	modified:   logic_building/8_struct/struct.c
	modified:   logic_building/finals/spiral_matrix_2.c
	modified:   logic_building/finals/tempCodeRunnerFile.c
	modified:   temp.c
	modified:   temp2.py
	new file:   tempCodeRunnerFile.c
	deleted:    tempCodeRunnerFile.py

+++ logic_building/1_basic/7_print_ascii.c
@@ -4,6 +4,6 @@ int main(void) {
     char ch;
     printf("Enter a character: ");
     scanf("%c", &ch);
-    printf("The ASCII value of '%c' is %d\n", ch, (int)ch);
+    printf("The ASCII value of '%c' is %d\n", ch, ch);
     return 0;
 }
